refactor: move files related to Scatter
