+++ giraffe/src/components/Scatter/index.ts
@@ -1,0 +1,1 @@
+export {ScatterLayer} from './ScatterLayer'
