+++ giraffe/src/components/SizedPlot.tsx
@@ -25,7 +25,7 @@ import {AnnotationLayer} from './Annotation'
 import {SingleStatLayer} from './SingleStatLayer'
 import {LineLayer} from './Line'
 import {BandLayer} from './Band'
-import {ScatterLayer} from './ScatterLayer'
+import {ScatterLayer} from './Scatter'
 import {RectLayer} from './RectLayer'
 import {MosaicLayer} from './Mosaic'
 import GeoLayer from './GeoLayer'
